requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
pydantic==2.12.5
pydantic-settings==2.12.0
sqlalchemy==2.0.45
psycopg2-binary==2.9.11
httpx==0.28.1
openai==2.15.0
twilio==9.9.1
pandas==2.3.3
openpyxl==3.1.5
logfire[fastapi]
pydantic-ai-slim[openai]
python-multipart
